Check Restaurant Inspection status based on their license number
Ask the user to input his license number to check whether it is valid
or invalid. If it is invalid, the restaurant should immediately renew
license
It shows the user the date the inspection was carried out from the most
recent one to the least recent one.
The comparison actually checks the number of Pass with the number of
other values.
If the number of pass is equal to the number of fail, then the license
number is invalid